% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censAr-package.R
\docType{package}
\name{censAr-package}
\alias{censAr}
\alias{censAr-package}
\title{censAr: Base de Datos de Facil Acceso de Censos de Arentina (Argentina´s Census Easy Access Database)}
\description{
Provee un acceso conveniente a mas de 40 millones de registros de la base de datos del Censo 2010 de Argentina. Los datos fueron obtenidos de 'datar', donde se publican datos en texto plano a partir del Convertidor REDATAM creado por Pablo De Grande.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://politicaargentina.github.io/censAr/}
  \item Report bugs at \url{https://github.com/politicaargentina/censAr/issues/}
}

}
\author{
\strong{Maintainer}: Juan Pablo Ruiz Nicolini \email{juanpabloruiznicolini@gmail.com} [copyright holder]

}
\keyword{internal}
